Feature Update
Added a feature for you to be able to use source trigger again, instead of all pages. If the advertiser uses a cookie consent pop-up that redirects to the website, it would cause the template to not work. If you disable the "Using All Page Trigger" checkbox now and use the source trigger instead. It should work correctly.